fixed prod build
